requirements.txt
Pillow
numpy
pycuda
opencv-python